fix input freezing
This fixes an issue where input would be frozen after moving the windows a lot. By moving event handling to a separate thread, it no longer has to wait on the panning functionality, notably the part where it has to read camera frames. Otherwise, it would have to process a bunch of ConfigureNotify events one frame at a time.